requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
packaging==24.2
pillow==11.0.0
pytube==10.6.1
pytubefix==8.8.1
requests==2.32.3
urllib3==2.2.3
